# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AffectLog Trustworthy AI Assessment Building Block (ALT-AI)
  description: API for dataset profiling, xAPI normalization, fairness metrics, model
    explainability, and EU AI Act / GDPR compliance exports. Aligned with D3.7 EDGE-Skills
    / Prometheus-X.
  contact:
    name: AffectLog
    url: https://github.com/roy-saurabh/edge_affectlog
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.1.0
paths:
...
  /v1/models/register:
    post:
      tags:
      - Models
      summary: Register a model
      operationId: register_model_v1_models_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelRegisterRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelRegisterResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}:
    get:
      tags:
      - Models
      summary: Get model metadata
      operationId: get_model_v1_models__model_id__get
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Model V1 Models  Model Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}/predict:
    post:
      tags:
      - Models
      summary: Run model prediction
      operationId: predict_v1_models__model_id__predict_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/{model_id}/explain:
    post:
      tags:
      - Models
      summary: Generate model explanations
      operationId: explain_v1_models__model_id__explain_post
      parameters:
      - name: model_id
        in: path
        required: true
        schema:
          type: string
          title: Model Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PredictRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExplainResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/models/compare:
    post:
      tags:
      - Models
      summary: Compare registered models
      operationId: compare_models_endpoint_v1_models_compare_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_compare_models_endpoint_v1_models_compare_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompareResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Body_compare_models_endpoint_v1_models_compare_post:
      properties:
        model_ids:
          items:
            type: string
          type: array
          title: Model Ids
        req:
          $ref: '#/components/schemas/PredictRequest'
      type: object
      required:
      - model_ids
      - req
      title: Body_compare_models_endpoint_v1_models_compare_post
    CompareResponse:
      properties:
        comparison:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Comparison
      type: object
      required:
      - comparison
      title: CompareResponse
...
    ExplainResponse:
      properties:
        model_id:
          type: string
          title: Model Id
        explanations:
          additionalProperties: true
          type: object
          title: Explanations
      type: object
      required:
      - model_id
      - explanations
      title: ExplainResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ModelRegisterRequest:
      properties:
        model_path:
          type: string
          title: Model Path
        adapter_type:
          type: string
          title: Adapter Type
          description: sklearn | onnx | torch | tensorflow | http | dummy
        model_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Model Id
        feature_names:
          items:
            type: string
          type: array
          title: Feature Names
          default: []
      type: object
      required:
      - model_path
      - adapter_type
      title: ModelRegisterRequest
    ModelRegisterResponse:
      properties:
        model_id:
          type: string
          title: Model Id
        adapter_type:
          type: string
          title: Adapter Type
        status:
          type: string
          title: Status
      type: object
      required:
      - model_id
      - adapter_type
      - status
      title: ModelRegisterResponse
...
    PredictRequest:
      properties:
        instances:
          items:
            items: {}
            type: array
          type: array
          title: Instances
      type: object
      required:
      - instances
      title: PredictRequest
      example:
        instances:
        - - 1.0
          - 2.0
          - 3.0
    PredictResponse:
      properties:
        model_id:
          type: string
          title: Model Id
        predictions:
          items: {}
          type: array
          title: Predictions
      type: object
      required:
      - model_id
      - predictions
      title: PredictResponse
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError